requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
fastapi==0.115.5
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watsonx_ai==1.1.24
idna==3.10
importlib_metadata==8.5.0
JayDeBeApi==1.2.3
jmespath==1.0.1
jpype1==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.8
langchain-community==0.3.8
langchain-core==0.3.21
langchain-ibm==0.3.3
langchain-text-splitters==0.3.2
langsmith==0.1.146
lomond==0.3.3
lxml==5.3.0
marshmallow==3.23.1
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.12
packaging==24.2
pandas==2.1.4
peewee==3.17.8
platformdirs==4.3.6
propcache==0.2.0
pyarrow==18.1.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.2
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.41.3
tabulate==0.9.0
tenacity==9.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
webencodings==0.5.1
yarl==1.18.0
yfinance==0.2.50
zipp==3.21.0
